Consolidate more python functions
