--- ush/config_utils.py
@@ -1,16 +1,0 @@
-#!/usr/bin/env python3
-
-"""
-Interface to configuration file management utilities.
-To see what it can do, run:
-
-.. code-block:: console
-    
-   ./config_utils --help
-
-"""
-
-from python_utils import cfg_main
-
-if __name__ == "__main__":
-    cfg_main()
