+++ ush/python_utils/__init__.py
@@ -1,5 +1,6 @@
 from .misc import uppercase, lowercase, find_pattern_in_str, find_pattern_in_file, dict_find
 from .check_for_preexist_dir_file import check_for_preexist_dir_file
+from .check_python_version import check_python_version
 from .check_var_valid_value import check_var_valid_value
 from .create_symlink_to_file import create_symlink_to_file
 from .define_macos_utilities import define_macos_utilities
